Fixed search by free text
git-svn-id: https://svn.apache.org/repos/asf/rave/branches/mongo@1401862 13f79535-47bb-0310-9956-ffa450edef68

+++ rave-components/rave-mongodb/src/main/java/org/apache/rave/portal/model/util/MongoDbModelUtil.java
@@ -29,6 +29,7 @@ public class MongoDbModelUtil {
     private MongoDbModelUtil(){}
 
     public static long generateId() {
-        return Math.abs(new Random().nextLong());
+        //Use INT as javascript has issues with truncating longs and we are moving model Ids to String
+        return Math.abs(new Random().nextInt());
     }
 }
